App/requirements.txt
async-generator==1.10
attrs==22.2.0
Automat==22.10.0
blinker==1.7.0
boto3==1.34.34
botocore==1.34.34
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorthief==0.2.1
constantly==23.10.4
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cycler==0.11.0
dnspython==2.5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.39.3
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.4
lxml==4.9.3
MarkupSafe==2.1.5
matplotlib==3.7.1
numpy==1.24.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
pandas==2.1.4
parsel==1.8.1
Pillow==9.5.0
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pygame==2.5.0
pymongo==4.6.1
pymongo-auth-aws==1.1.0
pyOpenSSL==23.3.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
s3transfer==0.10.0
Scrapy==2.11.0
selenium==4.8.3
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tkinterx==0.0.9
tldextract==5.1.1
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.2
Werkzeug==3.0.1
wsproto==1.2.0
zope.interface==6.1
